removed all files we plan on never running in esm world, updated package.json to include the 2 suites we will run. finished porting segments tests, abstracted reusable express helpers
