--- test/versioned/express-esm/.eslintrc.cjs
@@ -1,14 +1,0 @@
-/*
- * Copyright 2022 New Relic Corporation. All rights reserved.
- * SPDX-License-Identifier: Apache-2.0
- */
-
-module.exports = {
-  parserOptions: {
-    ecmaVersion: '2020'
-  },
-  rules: {
-    'node/no-unpublished-import': 'off',
-    'node/no-extraneous-import': 'off'
-  }
-}
