+++ test/versioned/express-esm/newrelic.cjs
@@ -1,7 +1,9 @@
 /*
- * Copyright 2020 New Relic Corporation. All rights reserved.
+ * Copyright 2022 New Relic Corporation. All rights reserved.
  * SPDX-License-Identifier: Apache-2.0
  */
+
+'use strict'
 
 exports.config = {
   app_name: ['My Application'],
